+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add Paid all option
 - Switched away from com.example.... so now android thinks this is a different app
 - Change delete users shortcut from longpress to swipe
+- Add widget for home screen
 
 1.0.1
 - Add option to delete users